BUG Wrong output for very large input arrays.
Closes #35.

+++ bottleneck/src/template/move/move_nanmean.py
@@ -23,7 +23,7 @@ floats['top'] = """
 def NAME_NDIMd_DTYPE_axisAXIS(np.ndarray[np.DTYPE_t, ndim=NDIM] a,
                                   int window):
     "Moving mean of NDIMd array of dtype=DTYPE along axis=AXIS ignoring NaNs."
-    cdef int count = 0
+    cdef Py_ssize_t count = 0
     cdef double asum = 0, aold, ai
 """
